+++ luqum/parser.py
@@ -10,6 +10,7 @@ import re
 import ply.lex as lex
 import ply.yacc as yacc
 
+from .head_tail import TokenValue, head_tail, token_headtail
 from .tree import *
 
 
@@ -42,19 +43,6 @@ tokens = (
      'RBRACKET'] +
     # we sort to have a deterministic order, so that gammar signature does not changes
     sorted(list(reserved.values())))
-
-
-# text of some simple tokens
-t_PLUS = r'\+'
-t_MINUS = r'\-'
-t_NOT = 'NOT'
-t_AND_OP = r'AND'
-t_OR_OP = r'OR'
-t_COLUMN = r':'
-t_LPAREN = r'\('
-t_RPAREN = r'\)'
-t_LBRACKET = r'(\[|\{)'
-t_RBRACKET = r'(\]|\})'
 
 
 # precedence rules
@@ -133,9 +121,60 @@ REGEX_RE = r'''
   /         # closing slash
 )'''
 
+
 def t_SEPARATOR(t):
     r'\s+'
-    pass  # discard separators
+    token_headtail(t)
+    return None # discard separators
+
+
+# standard function for simple text tokens
+def simple_token(t):
+    t.value = TokenValue(t.value)
+    token_headtail(t)
+    return t
+
+# text of some simple tokens
+def t_PLUS(t):
+    r'\+'
+    return simple_token(t)
+
+def t_MINUS(t):
+    r'\-'
+    return simple_token(t)
+
+def t_NOT(t):
+    'NOT'
+    return simple_token(t)
+
+def t_AND_OP(t):
+    'AND'
+    return simple_token(t)
+
+def t_OR_OP(t):
+    r'OR'
+    return simple_token(t)
+
+def t_COLUMN(t):
+    r':'
+    return simple_token(t)
+
+def t_LPAREN(t):
+    r'\('
+    return simple_token(t)
+
+def t_RPAREN(t):
+    r'\)'
+    return simple_token(t)
+
+def t_LBRACKET(t):
+    r'(\[|\{)'
+    return simple_token(t)
+
+def t_RBRACKET(t):
+    r'(\]|\})'
+    return simple_token(t)
+
 
 @lex.TOKEN(TERM_RE)
 def t_TERM(t):
@@ -146,6 +185,9 @@ def t_TERM(t):
         m = re.match(TERM_RE, t.value, re.VERBOSE)
         value = m.group("term")
         t.value = Word(value)
+    else:
+        t.value = TokenValue(t.value)  # gentle wrapper to hande pos, tail, head
+    token_headtail(t)
     return t
 
 
@@ -154,33 +196,38 @@ def t_PHRASE(t):
     m = re.match(PHRASE_RE, t.value, re.VERBOSE)
     value = m.group("phrase")
     t.value = Phrase(value)
+    token_headtail(t)
     return t
+
 
 @lex.TOKEN(REGEX_RE)
 def t_REGEX(t):
     m = re.match(REGEX_RE, t.value, re.VERBOSE)
     value = m.group("regex")
     t.value = Regex(value)
+    token_headtail(t)
     return t
 
 
 @lex.TOKEN(APPROX_RE)
 def t_APPROX(t):
     m = re.match(APPROX_RE, t.value)
-    t.value = m.group("degree")
+    t.value = TokenValue(m.group("degree"))
+    token_headtail(t)
     return t
 
 
 @lex.TOKEN(BOOST_RE)
 def t_BOOST(t):
     m = re.match(BOOST_RE, t.value)
-    t.value = m.group("force")
+    t.value = TokenValue(m.group("force"))
+    token_headtail(t)
     return t
 
 
 # Error handling rule FIXME
 def t_error(t):  # pragma: no cover
-    print("Illegal character '%s'" % t.value[0])
+    print("Illegal character '%s' at position %d" % (t.value[0], t.lexpos))
     t.lexer.skip(1)
 
 
@@ -189,32 +236,38 @@ lexer = lex.lex()
 
 def p_expression_or(p):
     'expression : expression OR_OP expression'
-    p[0] = create_operation(OrOperation, p[1], p[3])
+    p[0] = create_operation(OrOperation, p[1], p[3], op_tail=p[2].tail)
+    head_tail.pos(p, head_transfer=False)
 
 
 def p_expression_and(p):
     '''expression : expression AND_OP expression'''
-    p[0] = create_operation(AndOperation, p[1], p[len(p) - 1])
+    p[0] = create_operation(AndOperation, p[1], p[3], op_tail=p[2].tail)
+    head_tail.pos(p, head_transfer=False)
 
 
 def p_expression_implicit(p):
     '''expression : expression expression'''
-    p[0] = create_operation(UnknownOperation, p[1], p[2])
+    p[0] = create_operation(UnknownOperation, p[1], p[2], op_tail="")
+    head_tail.pos(p, head_transfer=False)
 
 
 def p_expression_plus(p):
     '''unary_expression : PLUS unary_expression'''
     p[0] = Plus(p[2])
+    head_tail.unary(p)
 
 
 def p_expression_minus(p):
     '''unary_expression : MINUS unary_expression'''
     p[0] = Prohibit(p[2])
+    head_tail.unary(p)
 
 
 def p_expression_not(p):
     '''unary_expression : NOT unary_expression'''
     p[0] = Not(p[2])
+    head_tail.unary(p)
 
 
 def p_expression_unary(p):
@@ -224,14 +277,16 @@ def p_expression_unary(p):
 
 def p_grouping(p):
     'unary_expression : LPAREN expression RPAREN'
-    p[0] = Group(p[2])  # Will p_field_search will transform as FieldGroup if necessary
+    p[0] = Group(p[2])  # p_field_search will transform to FieldGroup if necessary
+    head_tail.paren(p)
 
 
 def p_range(p):
     '''unary_expression : LBRACKET phrase_or_term TO phrase_or_term RBRACKET'''
-    include_low = p[1] == "["
-    include_high = p[5] == "]"
+    include_low = p[1].value == "["
+    include_high = p[5].value == "]"
     p[0] = Range(p[2], p[4], include_low, include_high)
+    head_tail.range(p)
 
 
 def p_field_search(p):
@@ -240,6 +295,7 @@ def p_field_search(p):
         p[3] = group_to_fieldgroup(p[3])
     # for field name we take p[1].value for it was captured as a word expression
     p[0] = SearchField(p[1].value, p[3])
+    head_tail.search_field(p)
 
 
 def p_quoting(p):
@@ -249,12 +305,14 @@ def p_quoting(p):
 
 def p_proximity(p):
     '''unary_expression : PHRASE APPROX'''
-    p[0] = Proximity(p[1], p[2])
+    p[0] = Proximity(p[1], p[2].value)
+    head_tail.post_unary(p)
 
 
 def p_boosting(p):
     '''expression : expression BOOST'''
-    p[0] = Boost(p[1], p[2])
+    p[0] = Boost(p[1], p[2].value)
+    head_tail.post_unary(p)
 
 
 def p_terms(p):
@@ -264,16 +322,20 @@ def p_terms(p):
 
 def p_fuzzy(p):
     '''unary_expression : TERM APPROX'''
-    p[0] = Fuzzy(p[1], p[2])
+    p[0] = Fuzzy(p[1], p[2].value)
+    head_tail.post_unary(p)
+
 
 def p_regex(p):
     '''unary_expression : REGEX'''
     p[0] = p[1]
 
+
 # handling a special case, TO is reserved only in range
 def p_to_as_term(p):
     '''unary_expression : TO'''
-    p[0] = Word(p[1])
+    p[0] = Word(p[1].value)
+    head_tail.simple_term(p)
 
 
 def p_phrase_or_term(p):